fixed save filter
