+++ src/Author/Controller/SavedFilterController.php
@@ -2,20 +2,14 @@
 
 namespace GisClient\Author\Controller;
 
+use GisClient\Author\Form\Type\FilterType;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Form\Forms;
-use Symfony\Component\Form\CallbackTransformer;
 use Symfony\Component\Form\Extension\HttpFoundation\HttpFoundationExtension;
 use Symfony\Component\Form\Extension\Validator\ValidatorExtension;
-use Symfony\Component\Form\Exception\TransformationFailedException;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Validator\Constraints\NotBlank;
-use Symfony\Component\Validator\Constraints\Required;
-use Symfony\Component\Validator\Constraints\Callback;
 use Symfony\Component\Validator\Validation;
 
 class SavedFilterController
@@ -31,153 +25,23 @@ class SavedFilterController
         return $formFactory;
     }
 
-    private function getForm()
+    private function getForm(array $data)
     {
         $formFactory = $this->getFormFactory();
-
-        $formBuilder = $formFactory->createBuilder(
-            'Symfony\\Component\\Form\\Extension\\Core\\Type\\FormType',
-            null,
-            [
-                'allow_extra_fields' => true,
-            ]
-        )
-        ->add('saved_filter_name', TextType::class, [
-            'constraints' => [new Required(), new NotBlank()],
-        ])
-        ->add('mapset_name', TextType::class, [
-            'constraints' => [new Required(), new NotBlank(), new Callback([
-                'callback' => function ($value, $context) {
-                    $db = \GCApp::getDB();
-                    $sql = "SELECT * FROM ".DB_SCHEMA.".mapset WHERE mapset_name=?";
-                    $stmt = $db->prepare($sql);
-                    $stmt->execute([$value]);
-                    $data = $stmt->fetch(\PDO::FETCH_ASSOC);
-                    if ($data === false) {
-                        $context->addViolation('The mapset does not exists!');
-                    }
-                }
-            ])],
-        ])
-        ->add('layer_id', TextType::class, [
-            'constraints' => [new Required(), new NotBlank(), new Callback([
-                'callback' => function ($value, $context) {
-                    $normData = $context->getRoot()->getNormData();
-
-                    $db = \GCApp::getDB();
-                    $sql = "
-                        SELECT * FROM ".DB_SCHEMA.".mapset
-                        INNER JOIN ".DB_SCHEMA.".mapset_layergroup USING(mapset_name)
-                        INNER JOIN ".DB_SCHEMA.".layer USING(layergroup_id)
-                        WHERE mapset_name=? AND layer_id=?
-                    ";
-                    $stmt = $db->prepare($sql);
-                    $stmt->execute([$normData['mapset_name'], $normData['layer_id']]);
-                    $data = $stmt->fetch(\PDO::FETCH_ASSOC);
-                    if ($data === false) {
-                        $context->addViolation('The layer is not enabled on this mapset!');
-                    }
-                }
-            ])],
-        ])
-        ->add('saved_filter_scope', TextType::class, [
-            'constraints' => [new Required(), new NotBlank()],
-        ])
-        ->add('saved_filter_data', TextType::class, [
-            'constraints' => [new Required(), new NotBlank(), new Callback([
-                'callback' => function ($value, $context) {
-                    if (is_string($value) && json_decode($value) === null) {
-                        $context->addViolation('Not a valid json.');
-                    }
-                }
-            ])],
-        ]);
-
-        $formBuilder->get('layer_id')
-            ->addViewTransformer(new CallbackTransformer(
-                function ($layerId) {
-                    $composedLayerName = null;
-                    if ($layerId !== null) {
-                        $db = \GCApp::getDB();
-                        $sql = "
-                            SELECT project_name||'.'||layergroup_name||'.'||layer_name FROM ".DB_SCHEMA.".theme
-                            INNER JOIN ".DB_SCHEMA.".layergroup USING(theme_id)
-                            INNER JOIN ".DB_SCHEMA.".layer USING(layergroup_id)
-                            WHERE layer_id=?
-                        ";
-                        $stmt = $db->prepare($sql);
-                        $stmt->execute([$layerId]);
-                        $composedLayerName = $stmt->fetchColumn();
-                        if ($composedLayerName === false) {
-                            throw new TransformationFailedException(sprintf(
-                                'No layer_id found for "%s"',
-                                $layerId
-                            ));
-                        }
-                    }
-                    return $composedLayerName;
-                },
-                function ($composedLayerName) {
-                    if (strpos($composedLayerName, '.') !== false) {
-                        list($projectName, $layergroupName, $layerName) = explode('.', $composedLayerName);
-                    } else {
-                        throw new TransformationFailedException(
-                            'The layer_id must provied layergroup and layername separated by dot.'
-                        );
-                    }
-
-                    $db = \GCApp::getDB();
-                    $sql = "
-                        SELECT layer_id FROM ".DB_SCHEMA.".theme
-                        INNER JOIN ".DB_SCHEMA.".layergroup USING(theme_id)
-                        INNER JOIN ".DB_SCHEMA.".layer USING(layergroup_id)
-                        WHERE project_name=? AND layergroup_name=? AND layer_name=?
-                    ";
-                    $stmt = $db->prepare($sql);
-                    $stmt->execute([$projectName, $layergroupName, $layerName]);
-                    $layerId = $stmt->fetchColumn();
-                    if ($layerId === false) {
-                        throw new TransformationFailedException(sprintf(
-                            'No layer_id found for "%s"',
-                            $composedLayerName
-                        ));
-                    }
-                    
-                    return $layerId;
-                }
-            ));
-
-            $formBuilder->get('saved_filter_data')
-            ->addViewTransformer(new CallbackTransformer(
-                function ($jsonString) {
-                    return $jsonString;
-                },
-                function ($json) {
-                    if (is_array($json)) {
-                        $jsonString = json_encode($json);
-                    } else {
-                        $jsonString = $json;
-                    }
-                    
-                    return $jsonString;
-                }
-            ));
-
-        return $formBuilder->getForm();
+        return $formFactory->create(FilterType::class, $data);
     }
 
-    private function validate(array $data, array $originalData = [])
+    private function validate(array $data, $originalData = [])
     {
         $errors = [];
 
-        $form = $this->getForm();
+        $form = $this->getForm($originalData);
         if (count($originalData) > 0) {
-            $form->setData($originalData);
             $form->submit($data, false);
         } else {
             $form->submit($data);
         }
-
+        
         // check if is valid
         if (!$form->isValid()) {
             foreach ($form->getErrors(true) as $error) {
